requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
pydantic==2.12.0
yt-dlp